slight refactor, show actions on selection
